change batch_idx=-1 from 0 when table [batch_info] has no record, then batch_idx++ and create_new_batch(), first vector file_no is 0 not 1

+++ IndexIVF_DB.h
@@ -71,7 +71,7 @@ class Index_DB {
     int SetSysConfig(system_conf_t &sys_conf);
     int GetSysConfig(system_conf_t &sys_conf);
 
-    int GetLatestBatch(size_t &batch);
+    int GetLatestBatch(int &batch);
 
   private:
     int DropTable(char *tbl_nm);
